requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
distro==1.8.0
dj-rest-auth==2.2.4
Django==4.2
django-allauth==0.58.2
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
numpy==1.26.2
oauthlib==3.2.2
openai==1.4.0
pandas==2.1.4
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
